Fix internal error with cast of readonly array to byte array

+++ bvm/ballerina-runtime/src/main/java/io/ballerina/runtime/internal/values/ArrayValueImpl.java
@@ -347,7 +347,7 @@ public class ArrayValueImpl extends AbstractArrayValue {
         if (byteValues != null) {
             return byteValues[(int) index];
         }
-        return ((Integer) refValues[(int) index]).byteValue();
+        return ((Long) refValues[(int) index]).byteValue();
     }
 
     /**
